+++ core/src/com/unciv/logic/civilization/diplomacy/CityStateFunctions.kt
@@ -51,10 +51,11 @@ class CityStateFunctions(val civInfo: Civilization) {
 
         // Unique unit for militaristic city-states
         if (uniqueTypes.contains(UniqueType.CityStateMilitaryUnits)) {
-            val possibleUnits = ruleset.units.values.filter { it.requiredTechs().any()
-                && ruleset.eras[ruleset.technologies[it.requiredTech!!]!!.era()]!!.eraNumber > ruleset.eras[startingEra]!!.eraNumber // Not from the start era or before
-                && it.uniqueTo != null && it.uniqueTo in unusedMajorCivs // Must be from a major civ not in the game
-                && ruleset.unitTypes[it.unitType]!!.isLandUnit() && ( it.strength > 0 || it.rangedStrength > 0 ) } // Must be a land military unit
+            val possibleUnits = ruleset.units.values.filter {
+                return@filter !it.availableInEra(ruleset, startingEra) // Not from the start era or before
+                    && it.uniqueTo != null && it.uniqueTo in unusedMajorCivs // Must be from a major civ not in the game
+                    && ruleset.unitTypes[it.unitType]!!.isLandUnit() && (it.strength > 0 || it.rangedStrength > 0) // Must be a land military unit
+            }
             if (possibleUnits.isNotEmpty())
                 civInfo.cityStateUniqueUnit = possibleUnits.random().name
         }
@@ -91,9 +92,9 @@ class CityStateFunctions(val civInfo: Civilization) {
         fun giftableUniqueUnit(): BaseUnit? {
             val uniqueUnit = civInfo.gameInfo.ruleset.units[civInfo.cityStateUniqueUnit]
                 ?: return null
-            if (uniqueUnit.requiredTech != null && !receivingCiv.tech.isResearched(uniqueUnit.requiredTech!!))
+            if (!receivingCiv.tech.isResearched(uniqueUnit))
                 return null
-            if (uniqueUnit.obsoleteTech != null && receivingCiv.tech.isResearched(uniqueUnit.obsoleteTech!!))
+            if (receivingCiv.tech.isObsolete(uniqueUnit))
                 return null
             return uniqueUnit
         }