+++ core/src/com/unciv/logic/map/tile/Tile.kt
@@ -27,6 +27,7 @@ import com.unciv.models.ruleset.unique.UniqueType
 import com.unciv.ui.components.extensions.withItem
 import com.unciv.ui.components.extensions.withoutItem
 import com.unciv.utils.DebugUtils
+import com.unciv.utils.Log
 import kotlin.math.abs
 import kotlin.math.min
 import kotlin.random.Random
@@ -305,8 +306,8 @@ open class Tile : IsPartOfGameInfoSerialization {
     fun removeImprovement() =
         improvementFunctions.changeImprovement(null)
 
-    fun changeImprovement(improvementStr: String, civToHandleCompletion:Civilization? = null) =
-        improvementFunctions.changeImprovement(improvementStr, civToHandleCompletion)
+    fun changeImprovement(improvementStr: String, civToHandleCompletion: Civilization? = null, unit: MapUnit? = null) =
+        improvementFunctions.changeImprovement(improvementStr, civToHandleCompletion, unit)
 
     // function handling when adding a road to the tile
     fun addRoad(roadType: RoadStatus, creatingCivInfo: Civilization?) {
@@ -670,7 +671,19 @@ open class Tile : IsPartOfGameInfoSerialization {
     }
 
     fun hasConnection(civInfo: Civilization) =
-            getUnpillagedRoad() != RoadStatus.None || forestOrJungleAreRoads(civInfo)
+        getUnpillagedRoad() != RoadStatus.None || forestOrJungleAreRoads(civInfo)
+
+    fun hasRoadConnection(civInfo: Civilization, mustBeUnpillaged: Boolean) =
+        if (mustBeUnpillaged)
+            (getUnpillagedRoad() == RoadStatus.Road) || forestOrJungleAreRoads(civInfo)
+        else
+            roadStatus == RoadStatus.Road || forestOrJungleAreRoads(civInfo)
+
+    fun hasRailroadConnection(mustBeUnpillaged: Boolean) =
+        if (mustBeUnpillaged)
+            getUnpillagedRoad() == RoadStatus.Railroad
+        else
+            roadStatus == RoadStatus.Railroad
 
 
     private fun forestOrJungleAreRoads(civInfo: Civilization) =
@@ -895,13 +908,25 @@ open class Tile : IsPartOfGameInfoSerialization {
             // otherwise use pillage/repair systems
             if (canPillageTileImprovement())
                 improvementIsPillaged = true
-            else
+            else {
                 roadIsPillaged = true
+                clearAllPathfindingCaches()
+            }
         }
 
         owningCity?.reassignPopulationDeferred()
         if (owningCity != null)
             owningCity!!.civ.cache.updateCivResources()
+    }
+
+    private fun clearAllPathfindingCaches() {
+        val units = tileMap.gameInfo.civilizations.asSequence()
+            .filter { it.isAlive() }
+            .flatMap { it.units.getCivUnits() }
+        Log.debug("%s: road pillaged, clearing cache for %d units", this, { units.count() })
+        for (otherUnit in units) {
+            otherUnit.movement.clearPathfindingCache()
+        }
     }
 
     fun isPillaged(): Boolean = improvementIsPillaged || roadIsPillaged