+++ core/src/com/unciv/logic/map/tile/TileInfoImprovementFunctions.kt
@@ -6,8 +6,10 @@ import com.unciv.logic.civilization.LocationAction
 import com.unciv.logic.civilization.NotificationCategory
 import com.unciv.logic.civilization.NotificationIcon
 import com.unciv.logic.civilization.diplomacy.DiplomaticModifiers
+import com.unciv.logic.map.mapunit.MapUnit
 import com.unciv.models.ruleset.tile.TileImprovement
 import com.unciv.models.ruleset.unique.StateForConditionals
+import com.unciv.models.ruleset.unique.UniqueTriggerActivation
 import com.unciv.models.ruleset.unique.UniqueType
 
 
@@ -187,7 +189,7 @@ class TileInfoImprovementFunctions(val tile: Tile) {
 
     fun changeImprovement(improvementName: String?,
                           /** For road assignment and taking over tiles - DO NOT pass when simulating improvement effects! */
-                          civToActivateBroaderEffects:Civilization? = null) {
+                          civToActivateBroaderEffects: Civilization? = null, unit: MapUnit? = null) {
         val improvementObject = tile.ruleset.tileImprovements[improvementName]
 
         when {
@@ -223,6 +225,9 @@ class TileInfoImprovementFunctions(val tile: Tile) {
         )
             takeOverTilesAround(civToActivateBroaderEffects, tile)
 
+        if (civToActivateBroaderEffects != null && improvementObject != null)
+            triggerImprovementUniques(improvementObject, civToActivateBroaderEffects, unit)
+
         val city = tile.owningCity
         if (city != null) {
             city.cityStats.update()
@@ -231,6 +236,29 @@ class TileInfoImprovementFunctions(val tile: Tile) {
                 city.reassignPopulationDeferred()
             }
         }
+    }
+
+    private fun triggerImprovementUniques(
+        improvement: TileImprovement,
+        civ: Civilization,
+        unit: MapUnit? = null
+    ) {
+        for (unique in improvement.uniqueObjects.filter { !it.hasTriggerConditional() })
+            if (unit != null) {
+                UniqueTriggerActivation.triggerUnitwideUnique(unique, unit)
+            }
+            else UniqueTriggerActivation.triggerCivwideUnique(unique, civ, tile = tile)
+
+        if (unit != null){
+            for (unique in unit.getTriggeredUniques(UniqueType.TriggerUponBuildingImprovement)
+                .filter { improvement.matchesFilter(it.params[0]) })
+                UniqueTriggerActivation.triggerUnitwideUnique(unique, unit)
+            }
+
+        for (unique in civ.getMatchingUniques(
+            UniqueType.TriggerUponBuildingImprovement, StateForConditionals(civInfo = civ, unit = unit))
+            .filter { improvement.matchesFilter(it.params[0]) })
+            UniqueTriggerActivation.triggerCivwideUnique(unique, civ, tile = tile)
     }
 
     private fun adtivateRemovalImprovement(