+++ core/src/com/unciv/models/ruleset/validation/RulesetValidator.kt
@@ -178,6 +178,7 @@ class RulesetValidator(val ruleset: Ruleset) {
         tryFixUnknownUniques: Boolean
     ) {
         for (reward in ruleset.ruinRewards.values) {
+            if (reward.weight < 0) lines += "${reward.name} has a negative weight, which is not allowed!"
             for (difficulty in reward.excludedDifficulties)
                 if (!ruleset.difficulties.containsKey(difficulty))
                     lines += "${reward.name} references difficulty ${difficulty}, which does not exist!"
@@ -657,25 +658,27 @@ class RulesetValidator(val ruleset: Ruleset) {
         for (requiredTech: String in unit.requiredTechs())
             if (!ruleset.technologies.containsKey(requiredTech))
                 lines += "${unit.name} requires tech ${requiredTech} which does not exist!"
-        if (unit.obsoleteTech != null && !ruleset.technologies.containsKey(unit.obsoleteTech!!))
-            lines += "${unit.name} obsoletes at tech ${unit.obsoleteTech} which does not exist!"
+        for (obsoleteTech: String in unit.techsAtWhichNoLongerAvailable())
+            if (!ruleset.technologies.containsKey(obsoleteTech))
+                lines += "${unit.name} obsoletes at tech ${obsoleteTech} which does not exist!"
         if (unit.upgradesTo != null && !ruleset.units.containsKey(unit.upgradesTo!!))
             lines += "${unit.name} upgrades to unit ${unit.upgradesTo} which does not exist!"
 
         // Check that we don't obsolete ourselves before we can upgrade
-        if (unit.upgradesTo!=null && ruleset.units.containsKey(unit.upgradesTo!!)
-            && unit.obsoleteTech!=null && ruleset.technologies.containsKey(unit.obsoleteTech!!)) {
-            val upgradedUnit = ruleset.units[unit.upgradesTo!!]!!
-            for (requiredTech: String in upgradedUnit.requiredTechs())
-                if (requiredTech != unit.obsoleteTech
-                    && !getPrereqTree(unit.obsoleteTech!!).contains(requiredTech)
-                )
-                    lines.add(
-                        "${unit.name} obsoletes at tech ${unit.obsoleteTech}," +
-                            " and therefore ${requiredTech} for its upgrade ${upgradedUnit.name} may not yet be researched!",
-                        RulesetErrorSeverity.Warning
+        for (obsoleteTech: String in unit.techsThatObsoleteThis())
+            if (unit.upgradesTo!=null && ruleset.units.containsKey(unit.upgradesTo!!)
+                && ruleset.technologies.containsKey(obsoleteTech)) {
+                val upgradedUnit = ruleset.units[unit.upgradesTo!!]!!
+                for (requiredTech: String in upgradedUnit.requiredTechs())
+                    if (requiredTech != obsoleteTech
+                        && !getPrereqTree(obsoleteTech).contains(requiredTech)
                     )
-        }
+                        lines.add(
+                            "${unit.name} obsoletes at tech ${obsoleteTech}," +
+                                " and therefore ${requiredTech} for its upgrade ${upgradedUnit.name} may not yet be researched!",
+                            RulesetErrorSeverity.Warning
+                        )
+            }
 
         for (resource in unit.getResourceRequirementsPerTurn(StateForConditionals.IgnoreConditionals).keys)
             if (!ruleset.tileResources.containsKey(resource))