+++ core/src/com/unciv/ui/screens/pickerscreens/GreatPersonPickerScreen.kt
@@ -14,6 +14,7 @@ class GreatPersonPickerScreen(val civInfo:Civilization) : PickerScreen() {
     private var theChosenOne: BaseUnit? = null
 
     init {
+        UncivGame.Current.settings.autoPlay.stopAutoPlay()
         closeButton.isVisible = false
         rightSideButton.setText("Choose a free great person".tr())