+++ core/src/com/unciv/ui/screens/worldscreen/mainmenu/WorldScreenMenuPopup.kt
@@ -1,5 +1,6 @@
 package com.unciv.ui.screens.worldscreen.mainmenu
 
+import com.unciv.UncivGame
 import com.unciv.ui.components.input.KeyboardBinding
 import com.unciv.ui.components.input.onLongPress
 import com.unciv.ui.popups.Popup
@@ -10,6 +11,7 @@ import com.unciv.ui.screens.worldscreen.WorldScreen
 
 class WorldScreenMenuPopup(val worldScreen: WorldScreen) : Popup(worldScreen, scrollable = Scrollability.All) {
     init {
+        UncivGame.Current.settings.autoPlay.stopAutoPlay()
         defaults().fillX()
 
         addButton("Main menu") {